Move livecheck_strategy into root directory
