+++ livecheck_strategy.rb
