+++ livecheck_strategy/apache.rb
