+++ livecheck_strategy/git.rb
