+++ livecheck_strategy/gnome.rb
