+++ livecheck_strategy/gnu.rb
