+++ livecheck_strategy/hackage.rb
