+++ livecheck_strategy/launchpad.rb
