+++ livecheck_strategy/npm.rb
